+++ src/OrcCommand/Command/WolfLauncher/WolfLauncher.h
@@ -200,6 +200,7 @@ private:
     void ReadLogConfiguration(const ConfigItem& configItem, bool hasConsoleConfigItem);
 
 private:
+    GUID m_guid;
     ConsoleConfiguration m_consoleConfiguration;
 
     Journal m_journal;