+++ src/OrcCommand/Command/WolfLauncher/WolfLauncher_Run.cpp
@@ -29,6 +29,7 @@
 #include "CryptoHashStream.h"
 
 #include "Utils/Guard.h"
+#include "Utils/Guid.h"
 #include "Utils/TypeTraits.h"
 #include "Utils/Time.h"
 #include "Utils/WinApi.h"
@@ -47,9 +48,10 @@ namespace {
 
 constexpr std::wstring_view kInfo = L"Info";
 
-struct FileInformations
+class FileInformations
 {
-    bool exist;
+public:
+    bool file_exists = false;
     std::wstring path;
     std::optional<Orc::Traits::ByteQuantity<uint64_t>> size;
 };
@@ -70,7 +72,7 @@ GetLocalOutputFileInformations(const Orc::Command::Wolf::WolfExecution& exec, Fi
     }
 
     fileInformations.path = exec.GetOutputFullPath();
-    fileInformations.exist = true;
+    fileInformations.file_exists = true;
 
     WIN32_FILE_ATTRIBUTE_DATA data;
     ZeroMemory(&data, sizeof(WIN32_FILE_ATTRIBUTE_DATA));
@@ -156,7 +158,13 @@ HRESULT GetRemoteOutputFileInformations(
             return hr;
         }
 
-        fileInformations.exist = true;
+        if (hr == S_FALSE)
+        {
+            fileInformations.file_exists = false;
+            return S_OK;
+        }
+
+        fileInformations.file_exists = true;
         fileInformations.path = uploadAgent.GetRemoteFullPath(exec.GetOutputFileName());
         fileInformations.size = fileSize;
         return S_OK;
@@ -203,9 +211,11 @@ Result<std::wstring> GetProcessExecutableHash(DWORD dwProcessId, CryptoHashStrea
     return GetProcessExecutableHash(*hProcess, algorithm);
 }
 
-void UpdateOutcome(Command::Wolf::Outcome::Outcome& outcome, HANDLE hMothership)
+void UpdateOutcome(Command::Wolf::Outcome::Outcome& outcome, const GUID& id, HANDLE hMothership)
 {
     auto&& lock = outcome.Lock();
+
+    outcome.SetId(id);
 
     {
         std::wstring computerName;
@@ -565,6 +575,10 @@ HRESULT Orc::Command::Wolf::Main::CreateAndUploadOutline()
         writer->BeginElement(L"dfir-orc");
         writer->BeginElement(L"outline");
         {
+            std::wstring id;
+            Orc::ToString(m_guid, std::back_inserter(id));
+            writer->WriteNamed(L"id", id);
+
             writer->WriteNamed(L"version", kOrcFileVerStringW);
 
             std::wstring start;
@@ -578,10 +592,14 @@ HRESULT Orc::Command::Wolf::Main::CreateAndUploadOutline()
             auto mothership_id = SystemDetails::GetParentProcessId();
             if (mothership_id)
             {
+                const wchar_t kMothership[] = L"mothership";
+
+                writer->BeginElement(kMothership);
+
                 auto mothership_cmdline = SystemDetails::GetCmdLine(mothership_id.value());
                 if (mothership_cmdline)
                 {
-                    writer->WriteNamed(L"command", mothership_cmdline.value().c_str());
+                    writer->WriteNamed(L"command_line", mothership_cmdline.value().c_str());
                 }
 
                 const auto sha1 = GetProcessExecutableHash(mothership_id.value(), CryptoHashStream::Algorithm::SHA1);
@@ -589,6 +607,8 @@ HRESULT Orc::Command::Wolf::Main::CreateAndUploadOutline()
                 {
                     writer->WriteNamed(L"sha1", sha1.value());
                 }
+
+                writer->EndElement(kMothership);
             }
             writer->WriteNamed(L"output", config.Output.Path.c_str());
             writer->WriteNamed(L"temp", config.TempWorkingDir.Path.c_str());
@@ -716,7 +736,7 @@ Orc::Result<void> Main::CreateAndUploadOutcome()
         return Success<void>();
     }
 
-    ::UpdateOutcome(m_outcome, m_hMothership);
+    ::UpdateOutcome(m_outcome, m_guid, m_hMothership);
     ::UpdateOutcome(m_outcome, config.m_Recipients);
     ::UpdateOutcome(m_outcome, m_standardOutput);
     ::UpdateOutcome(m_outcome, m_logging);
@@ -940,7 +960,7 @@ HRESULT Main::Run_Execute()
                         const auto path = m_pUploadAgent->GetRemoteFullPath(exec->GetOutputFileName());
                         Log::Error(L"Failed to check remote file status: '{}' [{}]", path, SystemError(hr));
                     }
-                    else if (!info.size || *info.size != 0)
+                    else if (info.file_exists && (!info.size.has_value() || info.size.value() > 0))
                     {
                         commandSetNode.Add(
                             "Skipping set because non-empty remote output file already exists: '{}' (size: {})",