+++ src/OrcLib/BITSAgent.cpp
@@ -760,11 +760,11 @@ HRESULT BITSAgent::CheckFileUpload(const std::wstring& strRemoteName, std::optio
             }
             break;
         default:
-            Log::Debug(L"Invalid BITS Mode {}", m_config.bitsMode);
+            Log::Critical(L"Invalid BITS Mode {}", m_config.bitsMode);
             break;
     }
 
-    return S_OK;
+    return hr;
 }
 
 BITSAgent::~BITSAgent()