+++ src/OrcLib/CompleteVolumeReader.h
@@ -36,9 +36,11 @@ protected:
     CDiskExtentVector m_Extents;
 
     HRESULT ParseBootSector();
-    HRESULT Read(CBinaryBuffer& data, ULONGLONG ullBytesToRead, ULONGLONG& ullBytesRead);
 
 private:
+    HRESULT Read(CBinaryBuffer& data, ULONGLONG ullBytesToRead, ULONGLONG& ullBytesRead) override;
+    HRESULT ReadUnaligned(ULONGLONG offset, CBinaryBuffer& data, ULONGLONG ullBytesToRead, ULONGLONG& ullBytesRead);
+
     concurrency::critical_section m_cs;
 };